wayland/dma-buf: Add support for setting the deadline for a dma-buf
The extracted sync_file of a buffer supports an IOCTL to hint to the
kernel at which point the work associated with the buffer and sync_file
is supposed to finish. The idea is that the kernel can use this to keep
clocks high and make better scheduling decisions.

This is especially useful because the outstanding work of a wayland
transaction is not connected all the way to a KMS commit and instead is
waited on, on the CPU side, so the kernel doesn't understand that we
want to show the work on the next refresh cycle.