wayland/transaction: Find transaction deadlines and tell the kernel
By looking at all surfaces of a transaction and all stage views the
surfaces are located on, we can figure out the earliest time that the
transaction could be used in an update. This time is then used as a
deadline hint for the kernel which should help with scheduling
decisions.