drm-buffer/gbm: Export fences of the buffer to a sync_file
When the DrmBuffer comes from a gbm_surface, there might be work
committed on the front buffer that gets locked. This commit exports
fences for this work to a sync_file.

KMS still relies on implicit sync but the sync_file will be used in the
next commit to set a deadline hint on the work.