meta/selection-source-memory: Add multiple mimetype support
MetaSelectionSourceMemory currently supports only a single mimetype
per selection source. This is in contrast to MetaSelectionSourceWayland
or MetaSelectionSourceX11 that support multiple mimetypes in a single
selection source. This commit modifies the implementation of
MetaSelectionSourceMemory to support multiple mimetype values via a new
constructor, `_new_multiple`. The old API is still supported to maintain
backwards compatibility.