clutter/snapshot: Add private pipeline node wrapper
I'm a little hesitant to expose CoglPipeline directly as part of the
ClutterSnapshot API, so keep it private for now, only meant to be
consumed by ClutterOffscreenEffect.

I think ClutterOffscreenEffect should be eventually removed, and the
offscreen management should be internal to the rendering process, but
that's a fight for another day.