test-runner: Wait for stage update using after-update signal
Instead of using a BEFORE_REDRAW MetaLater to wait for the stage update, use
the usual way that we use in tests and wait for the stage update finishing
by connecting to the "::after-update" signal. This should ensure that the
test_case_dispatch() function actually blocks until the whole update has
finished, rather than before the redraw happens.

Arguably the BEFORE_REDRAW later was long enough to cover most cases (as by
then events have been handled and Clutter has done any pending relayouts),
but still it seems cleaner to use the "after-udpate" signal for this.