test-client: Flush request queue before sending OK back to runner
When we send OK back to the runner, the runner continues execution, but the
test client only flushes its wayland request queue after that (and only when
reading synchronously).

Ensure that the queue is always fully flushed *before* we hand control back
to the runner, so that the general expectation inf tests that the client
command has actually fully ran is fulfilled.