test-runner: Flush our Wayland event queue to clients on potential changes
Turns out we're not flushing our Wayland event queue after executing a
command in the test-runner. This means the client might not see the event
immediately, and might already be in the middle of executing the next
test command when the events come in.

Simply do a flush of mutters Wayland event queue after all commands. While
this includes a few unnecessary ones where no events got queued, it doesn't
hurt there either.