test-runner: Look for window actors instead of windows in wait_for_effects
When calling gtk_window_hide() on a window, gtk destroys the surface right
away, and we mark our corresponding MetaWindow as unmanaging. Since the last
commit ("test-client: Flush request queue before sending OK back to
runner"), mutter now sees this wl_surface.destroy() request immediately
after a call to "hide".

When calling "wait_for_effects" right after "hide" (which we do in
cloned-windows and in suspend-state-timeout tests), this means
meta_test_client_find_window() in "wait_for_effects" fails to find the
window, and the test often fails.

The MetaWindow is still around (just unmanaging), and the MetaWindowActor is
still around and doing the effect, so we can use a different method to find
the window actor that's doing the effect. Introduce a new util function
meta_test_client_find_window_actor() that does the same as
meta_test_client_find_window(), but falls back to looking through the list
of window actors.