tests/runner: Don't wait until struts have been updated in add/set_strut
set_strut and add_strut are supposed to add struts to the workarea, and
mutter queues this workarea update into an idle so that all the struts
changes can be batched together. In the test we currently circumvent this
batching by waiting for the "workareas-changed" signal when calling
add/set_strut. This means that window are reconfigured more often than they
would be in real life, and can even cause windows to need two new
configurations (which is the reason why the strut-changes test is flaky
sometimes).

Remove the wait for the "workareas-changed" signal. If a test wants to wait
and block until the struct has actually been applied, it can call the newly
introduced `wait_for_signal ::workareas-changed` command after set/add_strut
instead.