test-client: Use wl_display_flush() to flush queue instead of _roundtrip()
Tests are asynchronous by default, mutter and the test client are both
running and sending wayland events/requests at the same time. Before we do a
stage update and mutter reads the incoming wayland requests from the client,
we need to ensure that the client has flushed all its requests over the
socket.

For Wayland, currently we do this by asynchronously calling the "sync"
command on the client, which internally triggers a call to
wl_display_roundtrip() in GDKs wayland backend. Calling
wl_display_roundtrip() does the trick just fine, but it also waits until the
server has acked those requests.

However, this behavior of wl_display_roundtrip() requiring a response from
the server is not great for the case where the server is frozen/blocked
while the client is doing work (something we want to make possible with the
next commits), as in that case the client would deadlock, waiting for the
response from the server forever.

So switch from wl_display_roundtrip() to wl_display_flush() by replacing the
clients "sync" command with "flush", and calling gdk_display_flush() instead
of _sync().