wayland: Implement the wl_fixes interface
Mesa for example uses a bunch of registry objects which linger around
without a way to destroy them. The wl_fixes interface provides a way to
get rid of them. In the future it might grow more functionality to work
around issues in the core protocol that can't be fixed otherwise.