color-device: Track color states and apply HDR/colorspace calibration
The color device now updates the white point via a LUT, sets the
colorspace and HDR metadata properties on monitors to get into the
required configuration.

It then also creates color states from the current configuration. For
now this is only sRGB if the monitor is in "default" mode and PQ+BT.2020
when the monitor is in "HDR" mode.

In the future, the color state in the default mode should come from the
EDID color primaries or an ICC profile and a "brightness" setting should
define the reference white level.