stage-view: Move logic from RendererNative to the StageView classes
The ClutterStageView is now capabable of creating offscreens itself
instead of relying on the creator to create one. This will become
necessary when the color state changes need a fp16 offscreen for linear
blending.

The MetaRendererView now takes a MetaColorDevice which is used to always
get the up-to-date ClutterColorStates to the stage view.