wayland/drm-lease: Check drm fd before binding devices
The DRM lease protocol doesn't specify a way to report issues opening a
non-master fd for a DRM node when the client binds to the DRM lease
device.

To mitigate this issue do not bind the device when it is not possible to
open a non-privileged DRM fd during device creation.

Fixes: 536c52bdb10c ("wayland/drm-lease: Advertise initial devices")