x11/window: Check surface rectangle size in surface_rect_to_client_rect
Subtract the borders only if the surface rectangle is at least as large
as the borders. Prevents

*** BUG ***
In pixman_region32_init_rect: Invalid rectangle passed
Set a breakpoint on '_pixman_log_error' to debug

being logged by a debug build of pixman due to negative rectangle
width/height. Not sure what would happen in this case with a release
build.