window: Guard make_fullscreen()
Not all windows can be made fullscreen, but we currently rely
on all callers to check the condition beforehand.

Do the same as commit 7ff1c04c8fc6 did for minimize(), and add
an explicit check in the function itself to catch when it is
called unexpectedly.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4052>