Don't crash in selection "owner changed" if saved mimetype is null
We may end up in a situation when there is no new owner but there is
saved clipboard without saved MIME type. In this case calling
meta_selection_source_memory_new() fails and, worse, dereferencing null
error pointer when logging this failure results in a crash.

Avoid this by not calling it in this case because it is perfectly
possible for saved clipboard to be set to a non-null pointer in
transfer_cb, yet for the saved MIME type to be null if the data is not
available in any of the supported MIME types.