onscreen/native: Provide damage rectangles to secondary GPUs in copy path
This provides damage rectangles to the secondary GPU in the copy path so
it, much like the primary GPU, can avoid updating parts of the framebuffer
that haven't been changed since the last frame(s). This works in tandem
with EGL_EXT_buffer_age to determine how old the contents of the current
framebuffer are (i.e. to know if there is anything to reuse from the last
frame).