onscreen/native: Prevent duplicate damage rectangles in copy path
Since we use the buffer age to create a union of damage rectangles from the
last n frames, we often end up with the same rectangle multiple times if an
application happens to be changing that is in the same place. This in turn
results in duplicate vertices being generated and uploaded to the GPU in
the blitting and painting paths.

Spend some time (seems to take around 15 microseconds on average on my
system) dropping duplicate rectangles using a GHashTable. This partially
pays itself back CPU-wise by having to copy fewer rectangles in
copy_shared_framebuffer_gpu, looping and calculating less in
paint_egl_image. GPU-wise it uploads less data to the GPU and makes the
shaders invoke less.